修复

+++ gameserver/src/main/java/com/game/gameserver/module/goods/dao/EquipMapper.java
@@ -23,5 +23,5 @@ public interface EquipMapper {
      * @param bagPack  背包
      * @return java.util.List<com.game.gameserver.module.goods.entity.EquipEntity>
      */
-    List<Equip> getEquipEntityList(int playerId,int bagPack);
+    List<Equip> getEquipEntityList(long playerId,int bagPack);
 }
